+++ src/main/java/com/marcusfromsweden/plantdoctor/controller/GrowingLocationController.java
@@ -1,4 +1,4 @@
-package com.marcusfromsweden.plantdoctor.controller.internal;
+package com.marcusfromsweden.plantdoctor.controller;
 
 import com.marcusfromsweden.plantdoctor.dto.GrowingLocationDTO;
 import com.marcusfromsweden.plantdoctor.service.GrowingLocationService;
@@ -13,7 +13,7 @@ import java.util.List;
 import java.util.Optional;
 
 @RestController
-@RequestMapping("/api/internal/growing-locations")
+@RequestMapping("/api/growing-locations")
 public class GrowingLocationController {
 
     private static final Logger log = LoggerFactory.getLogger(GrowingLocationController.class);
